organnize layout and add flash messages for errors

+++ app.py
@@ -20,102 +20,15 @@ app.secret_key = "dev_secret_key"
 ENTRY_VIEW_LIMIT = 5
 
 
+"""
+______________
+HELPER FUNCS
+____________
+"""
+
+
 def is_logged_in():
     return session.get("username")
-
-
-def require_login(func):
-    @wraps(func)
-    def decorated_func(*args, **kwargs):
-        if not is_logged_in():
-            session["protected_page_path"] = request.path
-            flash("You must be logged in to do that.")
-            return redirect(url_for("display_login"))
-
-        return func(*args, **kwargs)
-
-    return decorated_func
-
-
-@app.before_request
-def load_db():
-    g.storage = DatabasePersistence()
-
-
-@app.route("/register")
-def display_register():
-    return render_template("register.html")
-
-
-@app.route("/register", methods=["POST"])
-def register():
-    username = request.form.get("username").strip().lower()
-    password = request.form.get("password").strip()
-
-    if username and password:
-        if not g.storage.user_exists(username):
-            hashed_pw = hashpw(password.encode("utf-8"), gensalt())
-            # convert to str before storing in db
-            hashed_pw_str = hashed_pw.decode("utf-8")
-
-            # create user
-            g.storage.register_new_user(username, hashed_pw_str)
-            # TODO add flash
-            print("==> flash Success: user is registered ")
-            return render_template("login.html")
-
-        else:
-            error = f"user {username} already exists"
-
-    else:
-        error = "missing input"
-
-    print(f"==> flash error: {error}, please try again")
-    return render_template("register.html", username=username), 422
-
-
-@app.route("/login")
-def display_login():
-    return render_template("login.html")
-
-
-@app.route("/login", methods=["POST"])
-def login():
-    username = request.form.get("username").strip().lower()
-    password = request.form.get("password").strip()
-
-    # validate input
-    if username and password:
-        result = g.storage.user_exists(username)
-        if result:
-            hashed_pw = result[2].encode("utf-8")
-            b_pw = password.encode("utf-8")
-            if checkpw(b_pw, hashed_pw):
-                # store username in session to track logged in user
-                session["username"] = username
-                session["user_id"] = result[0]
-                print("==> FLash: Success user authenticated")
-
-                # if proceeding from login redirect go back to original path
-                protected_page = session.pop("protected_page_path", None)
-                if protected_page:
-                    return redirect(session["protected_page_path"])
-
-                # default
-                return redirect(url_for("view_entries"))
-
-    # error redirect to same page with flash message
-    print(f"==> flash ERror: Invalid credentials, please try again")
-    return render_template("login.html"), 404
-
-
-@app.route("/logout", methods=["POST"])
-@require_login
-def logout():
-    session.clear()
-    print("You have been signed out.")
-    # TODO flash
-    return redirect(url_for("login"))
 
 
 def get_valid_page_nums():
@@ -130,45 +43,6 @@ def get_valid_page_nums():
     return valid_page_nums
 
 
-@app.route("/view_entries/", defaults={"page_num": 0})
-@app.route("/view_entries/<int:page_num>")
-@require_login
-def view_entries(page_num):
-    try:
-        user_id = session["user_id"]
-        entries_info = g.storage.get_users_entries_ids_and_date(
-            user_id, page_num, ENTRY_VIEW_LIMIT
-        )
-        print(f"SUCCESS entries: {entries_info}")
-        valid_page_nums_list = get_valid_page_nums()
-        # TODO determin page numbers for views that are acceptable (total)
-        if page_num not in valid_page_nums_list:
-            raise ValueError("Error: Page out of entry range")
-        return render_template(
-            "view_entries.html",
-            entries_info=entries_info,
-            page_num=page_num,
-            valid_page_nums_list=valid_page_nums_list,
-        )
-    except ValueError as e:
-        print(f"ERROR: {e}")
-        return (
-            render_template(
-                "view_entries.html",
-                entries_info=entries_info,
-                page_num=0,
-                valid_page_nums_list=valid_page_nums_list,
-            ),
-            404,
-        )  # TODO when to use render_template vs redirect
-
-
-@app.route("/create_entry")
-@require_login
-def display_create_entry():
-    return render_template("create_entry.html")
-
-
 def validate_entry(entry_date, energy_level, mood_range):
     if not entry_date:
         raise ValueError("Missing a date")
@@ -179,92 +53,6 @@ def validate_entry(entry_date, energy_level, mood_range):
         raise ValueError("Missing a value for energy level")
     if not mood_range:
         raise ValueError("Missing a value for mood range")
-
-
-@app.route("/create_entry", methods=["POST"])
-@require_login
-def create_entry():
-    # TODO maybe add a valid_request_body_keys_exist like budget.py
-    entry_date = request.form.get("entry_date")
-    energy_level = request.form.get("energy_level")
-    mood_range = request.form.get("mood_range")
-    emotions_string = request.form.get("emotions")
-    reflection = request.form.get("reflection")
-
-    user_id = session["user_id"]
-    try:
-        validate_entry(entry_date, energy_level, mood_range)
-        entry_id = g.storage.create_new_entry(
-            user_id, entry_date, energy_level, mood_range, reflection
-        )
-        print(f"SUCCESS: entry_id = {entry_id[0]}")
-        if emotions_string:
-            print(f"TEST emotions: {emotions_string}")
-            g.storage.add_emotions(entry_id[0], emotions_string)
-        return redirect(f"/view_entry/{entry_id[0]}")
-    except ValueError as e:
-        print(f"flash ERROR: {e}")
-        return (
-            render_template(
-                "create_entry.html",
-                entry_date=entry_date,
-                energy_level=energy_level,
-                mood_range=mood_range,
-                reflection=reflection,
-                emotions=emotions_string,
-            ),
-            404,
-        )
-    except Exception as e:
-        print(f"flash ERROR: {e}")
-
-    # TODO if invalid input redirect to same page with flash message for each error
-    # TODO intepolate valid entries so they dont have to rewrite
-    return render_template("create_entry.html")
-
-
-@app.route("/view_entry/<int:entry_id>")
-@require_login
-def display_entry(entry_id):  # TODO
-    try:
-        id, user_id, date, energy_level, mood_range, reflection = g.storage.get_entry(
-            entry_id
-        )
-        emotions_list = g.storage.get_entry_emotions(entry_id)
-        return render_template(
-            "view_entry.html",
-            entry_id=id,
-            date=date,
-            energy_level=energy_level,
-            mood_range=mood_range,
-            reflection=reflection,
-            emotions=emotions_list,
-        )
-    except TypeError:
-        print("ERROR: unauthorized entry")  # TODO flash
-        return render_template("view_entries.html"), 404
-
-
-@app.route("/edit_entry/<int:entry_id>")
-@require_login
-def display_edit_entry(entry_id):
-    try:
-        id, user_id, date, energy_level, mood_range, reflection = g.storage.get_entry(
-            entry_id
-        )
-        emotions_list = g.storage.get_entry_emotions(entry_id)
-        return render_template(
-            "edit_entry.html",
-            entry_id=entry_id,
-            entry_date=date,
-            energy_level=energy_level,
-            mood_range=mood_range,
-            reflection=reflection,
-            emotions=" ".join(emotions_list),
-        )
-    except TypeError:
-        print(f"ERROR: unauthorized entry")  # TODO flash
-        return render_template("view_entries.html"), 404
 
 
 def validate_edited_entry(entry_id, entry_date, energy_level, mood_range):
@@ -291,6 +79,225 @@ def validate_unique_emotions(emotions_string):
             count_emotions[emotion] = 1
 
 
+def require_login(func):
+    @wraps(func)
+    def decorated_func(*args, **kwargs):
+        if not is_logged_in():
+            session["protected_page_path"] = request.path
+            flash("You must be logged in to do that.")
+            return redirect(url_for("display_login"))
+
+        return func(*args, **kwargs)
+
+    return decorated_func
+
+
+@app.before_request
+def load_db():
+    g.storage = DatabasePersistence()
+
+
+"""
+____________
+ROUTES
+____________
+"""
+
+
+@app.route("/register")
+def display_register():
+    return render_template("register.html")
+
+
+@app.route("/register", methods=["POST"])
+def register():
+    username = request.form.get("username").strip().lower()
+    password = request.form.get("password").strip()
+
+    if username and password:
+        if not g.storage.user_exists(username):
+            hashed_pw = hashpw(password.encode("utf-8"), gensalt())
+            # convert to str before storing in db
+            hashed_pw_str = hashed_pw.decode("utf-8")
+
+            # create user
+            g.storage.register_new_user(username, hashed_pw_str)
+            flash(f"User '{username}', was registered", "success")
+            return render_template("login.html")
+
+        else:
+            error = f"User '{username}' already exists"
+
+    else:
+        error = "Missing input"
+
+    flash(f"{error}, please try again", "error")
+    return render_template("register.html", username=username), 422
+
+
+@app.route("/login")
+def display_login():
+    return render_template("login.html")
+
+
+@app.route("/login", methods=["POST"])
+def login():
+    username = request.form.get("username").strip().lower()
+    password = request.form.get("password").strip()
+
+    # validate input
+    if username and password:
+        result = g.storage.user_exists(username)
+        if result:
+            hashed_pw = result[2].encode("utf-8")
+            b_pw = password.encode("utf-8")
+            if checkpw(b_pw, hashed_pw):
+                # store username in session to track logged in user
+                session["username"] = username
+                session["user_id"] = result[0]
+
+                # if proceeding from login redirect go back to original path
+                protected_page = session.pop("protected_page_path", None)
+                if protected_page:
+                    return redirect(session["protected_page_path"])
+
+                # default
+                return redirect(url_for("view_entries"))
+
+    # error redirect to same page with flash message
+    flash("Invalid credentials, please try again", "error")
+    return render_template("login.html"), 404
+
+
+@app.route("/logout", methods=["POST"])
+@require_login  # TODO is this neccessary?
+def logout():
+    session.clear()
+    flash("You have been signed out.")
+    return redirect(url_for("login"))
+
+
+@app.route("/view_entries/", defaults={"page_num": 0})
+@app.route("/view_entries/<int:page_num>")
+@require_login
+def view_entries(page_num):
+    try:
+        user_id = session["user_id"]
+        entries_info = g.storage.get_users_entries_ids_and_date(
+            user_id, page_num, ENTRY_VIEW_LIMIT
+        )
+        valid_page_nums_list = get_valid_page_nums()
+        if page_num not in valid_page_nums_list:
+            raise ValueError("Error: Page out of entry range")
+        return render_template(
+            "view_entries.html",
+            entries_info=entries_info,
+            page_num=page_num,
+            valid_page_nums_list=valid_page_nums_list,
+        )
+    except ValueError as e:
+        flash(f"{e}", "error")
+        return (
+            render_template(
+                "view_entries.html",
+                entries_info=entries_info,
+                page_num=0,
+                valid_page_nums_list=valid_page_nums_list,
+            ),
+            404,
+        )  # TODO when to use render_template vs redirect
+
+
+@app.route("/create_entry")
+@require_login
+def display_create_entry():
+    return render_template("create_entry.html")
+
+
+@app.route("/create_entry", methods=["POST"])
+@require_login
+def create_entry():
+    # TODO maybe add a valid_request_body_keys_exist like budget.py
+    entry_date = request.form.get("entry_date")
+    energy_level = request.form.get("energy_level")
+    mood_range = request.form.get("mood_range")
+    emotions_string = request.form.get("emotions")
+    reflection = request.form.get("reflection")
+
+    user_id = session["user_id"]
+    try:
+        validate_entry(entry_date, energy_level, mood_range)
+        entry_id = g.storage.create_new_entry(
+            user_id, entry_date, energy_level, mood_range, reflection
+        )
+        if emotions_string:
+            g.storage.add_emotions(entry_id[0], emotions_string)
+        return redirect(f"/view_entry/{entry_id[0]}")
+    except ValueError as e:
+        flash(e, "error")
+        return (
+            render_template(
+                "create_entry.html",
+                entry_date=entry_date,
+                energy_level=energy_level,
+                mood_range=mood_range,
+                reflection=reflection,
+                emotions=emotions_string,
+            ),
+            404,
+        )
+    except Exception as e:
+        flash(e, "error")
+
+    # TODO if invalid input redirect to same page with flash message for each error
+    # TODO intepolate valid entries so they dont have to rewrite
+    return render_template("create_entry.html"), 504
+
+
+@app.route("/view_entry/<int:entry_id>")
+@require_login
+def display_entry(entry_id):  # TODO
+    try:
+        id, user_id, date, energy_level, mood_range, reflection = g.storage.get_entry(
+            entry_id
+        )
+        emotions_list = g.storage.get_entry_emotions(entry_id)
+        return render_template(
+            "view_entry.html",
+            entry_id=id,
+            date=date,
+            energy_level=energy_level,
+            mood_range=mood_range,
+            reflection=reflection,
+            emotions=emotions_list,
+        )
+    except TypeError:
+        flash("Unauthorized entry id", "error")
+        return redirect(url_for("view_entries")), 404
+
+
+@app.route("/edit_entry/<int:entry_id>")
+@require_login
+def display_edit_entry(entry_id):
+    try:
+        id, user_id, date, energy_level, mood_range, reflection = g.storage.get_entry(
+            entry_id
+        )
+        emotions_list = g.storage.get_entry_emotions(entry_id)
+        return render_template(
+            "edit_entry.html",
+            entry_id=entry_id,
+            entry_date=date,
+            energy_level=energy_level,
+            mood_range=mood_range,
+            reflection=reflection,
+            emotions=" ".join(emotions_list),
+        )
+    except TypeError:
+        flash("Unauthorized entry id", "error")
+        return redirect(url_for("view_entries")), 404
+
+
 @app.route("/edit_entry/<int:entry_id>", methods=["POST"])
 @require_login
 def edit_entry_and_emotions(entry_id):
@@ -313,7 +320,7 @@ def edit_entry_and_emotions(entry_id):
             g.storage.add_emotions(entry_id, emotions_string)
         return redirect(f"/view_entry/{entry_id}")
     except ValueError as e:
-        print(f"flash ERROR: {e}")
+        flash(e, "error")
         return (
             render_template(
                 "edit_entry.html",
@@ -327,15 +334,15 @@ def edit_entry_and_emotions(entry_id):
             404,
         )
     except Exception as e:
-        print(f"flash ERROR: {e}")
-        return render_template("view_entries.html"), 504
+        flash(e, "error")
+        return redirect(url_for("view_entries")), 504
 
 
 @app.route("/delete_entry/<int:entry_id>", methods=["POST"])
 @require_login
 def delete_entry(entry_id):
     g.storage.delete_entry(entry_id)
-    print("Successfully deleted entry...")
+    flash("Successfully deleted entry...", "success")  # TODO should i display date?
     return redirect(url_for("view_entries"))
 
 
@@ -347,6 +354,8 @@ def delete_entry(entry_id):
 def delete_emotion(entry_id, emotion):
     try:
         emotion_id = g.storage.get_emotions_id(emotion, entry_id)
+        if emotion_id is None:
+            raise TypeError()
         g.storage.delete_emotion(emotion_id, entry_id)
         id, user_id, date, energy_level, mood_range, reflection = g.storage.get_entry(
             entry_id
@@ -362,8 +371,8 @@ def delete_emotion(entry_id, emotion):
             emotions=emotions_list,
         )
     except TypeError:
-        print(f"ERROR: unauthorized emotion")  # TODO flash
-        return render_template("view_entries.html")
+        flash(f"Invalid emotion: {emotion}", "error")
+        return redirect(f"/edit_entry/{entry_id}"), 404
 
 
 if __name__ == "__main__":
